backend/requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
ale-py==0.7.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.5.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
AutoROM==0.6.1
AutoROM.accept-rom-license==0.6.1
backcall==0.2.0
bigjson==1.0.9
boto3==1.26.41
botocore==1.29.41
cachetools==4.2.1
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==2.1.1
clang==5.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.4
contourpy==1.0.7
cramjam==2.6.2
cycler==0.11.0
datasets==2.8.0
decorator==5.1.0
dill==0.3.6
distro==1.9.0
dm-tree==0.1.8
et-xmlfile==1.1.0
etils==0.9.0
exceptiongroup==1.3.1
fastapi==0.124.4
fastparquet==2022.12.0
filelock==3.8.2
flatbuffers==22.12.6
fonttools==4.39.2
frozenlist==1.3.3
fsspec==2022.11.0
gast==0.4.0
gin-config==0.1.3
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.56.4
grpcio==1.34.1
gym==0.21.0
gymnasium==0.27.1
gymnasium-notices==0.0.1
h11==0.16.0
h5py==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==4.13.0
importlib-resources==5.10.0
ipython==7.28.0
jax-jumpy==1.0.0
jedi==0.18.0
jiter==0.9.1
jmespath==1.0.1
joblib==1.1.0
keras==2.10.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keras-tuner==1.0.4
keyboard==0.13.5
kiwisolver==1.4.4
kt-legacy==1.0.4
libclang==14.0.6
luxai-s2==2.1.9
Markdown==3.3.3
markdown-it-py==2.2.0
matplotlib==3.7.1
matplotlib-inline==0.1.3
mdurl==0.1.2
multidict==6.0.3
multiprocess==0.70.14
noise==1.2.2
numpy==1.24.1
oauthlib==3.1.0
openai==2.2.0
opencv-python==4.7.0.72
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.0
pandas==1.5.2
parso==0.8.2
PettingZoo==1.12.0
pickleshare==0.7.5
Pillow==9.3.0
polars==0.15.11
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.15.2
psutil==5.9.4
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.3.0
pyglet==2.0.5
Pygments==2.15.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.1
pytorch-transformers==1.2.0
pytz==2021.3
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.0
responses==0.18.0
rich==13.3.4
rsa==4.7.2
s3transfer==0.6.0
sacremoses==0.0.53
scikit-learn==1.0
scipy==1.7.1
seaborn==0.12.2
sentencepiece==0.1.97
six==1.15.0
sklearn==0.0
sniffio==1.3.1
stable-baselines3==1.7.0
starlette==0.44.0
tb-nightly==2.5.0a20210224
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.10.1
tensorflow-addons==0.19.0
tensorflow-datasets==4.7.0
tensorflow-estimator==2.10.0
tensorflow-hub==0.12.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
tensorflow-metadata==1.11.0
tensorflow-probability==0.19.0
tensorflow-recommenders==0.7.2
tensorflow-text==2.10.0
termcolor==1.1.0
tf-agents==0.5.0
tf-estimator-nightly==2.5.0.dev2021022401
tf-nightly-gpu==2.5.0.dev20210223
threadpoolctl==3.0.0
tokenizers==0.13.2
toml==0.10.2
torch==1.11.0+cu115
torchaudio==0.11.0+cu115
torchvision==0.12.0+cu115
tqdm==4.64.1
traitlets==5.1.0
transformers==4.25.1
typeguard==2.13.3
typing_extensions==4.13.2
urllib3==1.26.13
uvicorn==0.33.0
vec-noise==1.1.4
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
xxhash==3.1.0
yarl==1.8.2
zipp==3.10.0
